# Update LinkedHash{Map,Set} from factory method