A fix to scala/bug#12766. I am running some benchmark to see if it can improve performance.