@liang3zy22 2023-04-13
I have run some benchmark using from factory method. The data are as below:
```
new from factory method benchmark:
[info] Benchmark                             (size)  (stringsOnly)  Mode  Cnt       Score      Error  Units
[info] LinkedHashMapBenchmark2.javalhmBuild      10           true  avgt   40     181.809 ±    1.107  ns/op
[info] LinkedHashMapBenchmark2.javalhmBuild     100           true  avgt   40    1639.742 ±    8.383  ns/op
[info] LinkedHashMapBenchmark2.javalhmBuild    1000           true  avgt   40   12484.748 ±   40.269  ns/op
[info] LinkedHashMapBenchmark2.javalhmBuild   10000           true  avgt   40  136926.001 ± 2104.757  ns/op
[info] LinkedHashMapBenchmark2.javalhsBuild      10           true  avgt   40     180.344 ±    0.631  ns/op
[info] LinkedHashMapBenchmark2.javalhsBuild     100           true  avgt   40    1560.707 ±   54.992  ns/op
[info] LinkedHashMapBenchmark2.javalhsBuild    1000           true  avgt   40   15311.915 ±  917.203  ns/op
[info] LinkedHashMapBenchmark2.javalhsBuild   10000           true  avgt   40  181002.984 ± 1331.890  ns/op
[info] LinkedHashMapBenchmark2.lhmBuild          10           true  avgt   40     161.782 ±    1.178  ns/op
[info] LinkedHashMapBenchmark2.lhmBuild         100           true  avgt   40    1475.403 ±    8.753  ns/op
[info] LinkedHashMapBenchmark2.lhmBuild        1000           true  avgt   40   14331.721 ±   56.220  ns/op
[info] LinkedHashMapBenchmark2.lhmBuild       10000           true  avgt   40  152435.558 ± 1095.835  ns/op
[info] LinkedHashMapBenchmark2.lhsBuild          10           true  avgt   40     221.315 ±    0.616  ns/op
[info] LinkedHashMapBenchmark2.lhsBuild         100           true  avgt   40    2220.355 ±    8.198  ns/op
[info] LinkedHashMapBenchmark2.lhsBuild        1000           true  avgt   40    9118.232 ±   32.174  ns/op
[info] LinkedHashMapBenchmark2.lhsBuild       10000           true  avgt   40  105495.135 ±  634.365  ns/op

old from factory method benchmark
[info] Benchmark                         (size)  (stringsOnly)  Mode  Cnt       Score      Error  Units
[info] LinkedHashMapBenchmark2.lhmBuild      10           true  avgt   40     164.116 ±    1.196  ns/op
[info] LinkedHashMapBenchmark2.lhmBuild     100           true  avgt   40    2847.231 ±   10.798  ns/op
[info] LinkedHashMapBenchmark2.lhmBuild    1000           true  avgt   40   20865.622 ±  117.579  ns/op
[info] LinkedHashMapBenchmark2.lhmBuild   10000           true  avgt   40  227393.939 ± 2353.514  ns/op
[info] LinkedHashMapBenchmark2.lhsBuild      10           true  avgt   40     150.682 ±    0.488  ns/op
[info] LinkedHashMapBenchmark2.lhsBuild     100           true  avgt   40    2352.999 ±   12.208  ns/op
[info] LinkedHashMapBenchmark2.lhsBuild    1000           true  avgt   40   17150.265 ±   80.456  ns/op
[info] LinkedHashMapBenchmark2.lhsBuild   10000           true  avgt   40  191911.733 ± 1985.895  ns/op


``