@lrytz 2023-04-13
Thank you @liang3zy22. Could you add unit tests to LinkedHashMapTest.scala / LinkedHashSetTest.scala?